requirements.txt
# =============================================================================
# Sen2Greenery - Dependencies
# =============================================================================
# Install these packages before running the app:
# pip install -r requirements.txt
# =============================================================================

# Core Streamlit framework (latest for best performance)
streamlit>=1.32.0

# Google Earth Engine integration
geemap>=0.30.0
earthengine-api>=0.1.380

# Map visualization
folium>=0.15.0
streamlit-folium>=0.15.0
geopandas>=0.14.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Image processing
Pillow>=10.0.0
matplotlib>=3.7.0

# Optional: for GeoTIFF support
rasterio>=1.3.0

# Performance optimization
cachetools>=5.3.0

